+++ Gruntfile.js
@@ -1,6 +1,13 @@
 // jshint maxlen:120
 'use strict';
 
+var LIVERELOAD_PORT = 35730;
+var SERVER_PORT = 9010;
+var lrSnippet = require('connect-livereload')({port: LIVERELOAD_PORT});
+var mountFolder = function (connect, dir) {
+  return connect.static(require('path').resolve(dir));
+};
+
 module.exports = function (grunt) {
   // load all grunt tasks
   require('load-grunt-tasks')(grunt);
@@ -10,9 +17,19 @@ module.exports = function (grunt) {
   grunt.initConfig({
     pkg: pkg,
     watch: {
+      options: {
+        nospawn: true,
+        livereload: LIVERELOAD_PORT
+      },
       yuidoc: {
         files: ['scripts/*.js'],
         tasks: ['clean', 'yuidoc']
+      },
+      livereload: {
+        files: [
+          'scripts/*.js',
+          'test/spec/*.js'
+        ]
       }
     },
     clean: {
@@ -76,8 +93,65 @@ module.exports = function (grunt) {
           src: ['dist/lateralus.js', 'dist/lateralus.min.js']
         }
       }
-    }
+    },
+    open: {
+      debug: {
+        path: 'http://localhost:' + SERVER_PORT + '/test'
+      }
+    },
+    connect: {
+      options: {
+        port: SERVER_PORT,
+        // change this to '0.0.0.0' to access the server from outside
+        hostname: 'localhost'
+      },
+      livereload: {
+        options: {
+          middleware: function (connect) {
+            return [
+              lrSnippet,
+              mountFolder(connect, './')
+            ];
+          }
+        }
+      }
+    },
+    /* jshint camelcase:false */
+    mocha_require_phantom: {
+      options: {
+        base: '.',
+        main: 'test/main',
+        requireLib: 'bower_components/requirejs/require.js',
+        files: ['test/spec/*.js']
+      },
+      debug: {
+        options: {
+          keepAlive: true
+        }
+      },
+      auto: {
+        options: {
+          keepAlive: false
+        }
+      }
+    },
   });
 
-  grunt.registerTask('default', ['clean', 'yuidoc', 'requirejs', 'usebanner']);
+  grunt.registerTask('test', [
+    'mocha_require_phantom:auto',
+  ]);
+
+  grunt.registerTask('debug', [
+    'open:debug',
+    'connect:livereload',
+    'watch:livereload'
+  ]);
+
+  grunt.registerTask('default', [
+    'test',
+    'clean',
+    'yuidoc',
+    'requirejs',
+    'usebanner'
+  ]);
 };