+++ scripts/lateralus.model.js
@@ -26,15 +26,16 @@ define([
    * @private
    * @class Lateralus.Model
    * @param {Lateralus} lateralus
+   * @param {Object} [options]
    * @extends {Backbone.Model}
    * @uses Lateralus.mixins
    * @constructor
    */
-  fn.constructor = function (lateralus) {
+  fn.constructor = function (lateralus, options) {
     this.lateralus = lateralus;
     this.delegateLateralusEvents();
     this.on('change', _.bind(this.onChange, this));
-    Backbone.Model.call(this);
+    Backbone.Model.call(this, options);
   };
 
   /**
@@ -46,11 +47,17 @@ define([
    * @method onChange
    */
   fn.onChange = function () {
-    var changed = this.changed;
+    var changed = this.changedAttributes();
 
     _.each(_.keys(changed), function (changedKey) {
       this.emit('change:' + changedKey, changed[changedKey]);
+
+      // Delete this property from the internal "changed" object before
+      // Backbone typically would to prevent "stacking" changed properties
+      // across onChange calls, thereby causing redundant handler calls.
+      delete this.changed[changedKey];
     }, this);
+
   };
 
   _.extend(fn, mixins);