+++ changelog.md
@@ -1,3 +1,9 @@
+# 0.5.0
+
+  * Add Lateralus.Router.
+  * Set up testing infrastructure and some tests/
+  * Prevent redundant global model change events.
+
 # 0.4.0
 
   * Emit events for all Lateralus.Model attribute changes.